chore: rename shopify-graphql.js => shopify.js
